resolve comments - rename file
