+++ lib/src/presentation/apod_full_size/apod_full_size_content.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:hello_universe/src/presentation/components/fade_in_network_image.dart';
+import 'package:hello_universe/src/presentation/components/apod_image.dart';
 
 class ApodFullSizeContent extends StatelessWidget {
   const ApodFullSizeContent({
